conditionally fetch bank transactions
